Move vote processing loop to VoteWorker
